move translations mocks to app-review
