+++ packages/@coorpacademy-app-review/src/views/skills/test/skills.test.ts
@@ -5,7 +5,7 @@ import omit from 'lodash/fp/omit';
 import {StoreState} from '../../../reducers';
 import {mapStateToSkillsProps} from '..';
 import {skin} from '../../slides/test/fixtures/skin';
-import {translate} from '../../../../sandbox/translation.mock';
+import {translate} from '../../../test/utils/translation.mock';
 
 const connectedOptions = {translate, onQuitClick: identity, skin};
